pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "microsoft-teams-cli"
version = "0.1.0"
description = "Microsoft Teams CLI — chat, send, and manage Teams from the terminal"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Yusuf Altunbicak"},
]
keywords = ["teams", "microsoft-teams", "chat", "cli", "office365", "messaging"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Office/Business",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "PyYAML>=6.0",
    "httpx>=0.25",
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3",
    "pytest-mock>=3.14",
    "respx>=0.22",
]

[project.urls]
Homepage = "https://github.com/yusufaltunbicak/microsoft-teams-cli"
Repository = "https://github.com/yusufaltunbicak/microsoft-teams-cli"
Issues = "https://github.com/yusufaltunbicak/microsoft-teams-cli/issues"

[project.scripts]
teams = "teams_cli.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["teams_cli"]

[tool.pytest.ini_options]
addopts = "-m 'not live'"
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "live: opt-in smoke tests against a real Teams account",
]
filterwarnings = [
    "ignore::bs4.MarkupResemblesLocatorWarning",
]
